Automated Boost linkage
The run_cpp procedure now searches for the names of the Boost linked libraries in the location provided by the user (and which is controlled by the variable boostdir). That allows linking with newer than the 1.65 version of the Boost libraries, which may have added or removed linked libraries, making the "hardwired" previous version of the procedure to fail.