Fixed "cleaning" annoyance
Added a new STATUS variable in which the status of the compilation command is being saved.
The value of this variable is now being checked before clearing the screen.
As a consequence, the screen is being cleared only on succesfull compilation.
This has the effect that more than one screenfull of errors can now be seen at the console,
making debugging programs much easier.